requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
certifi==2024.7.4
click==8.1.7
distro==1.9.0
Django==5.1.2
djangorestframework==3.15.2
dnspython==2.6.1
docutils==0.21.2
exceptiongroup==1.2.2
fastapi==0.112.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
jiter==0.5.0
numpy==2.0.1
openai==1.40.3
packaging==24.1
pluggy==1.5.0
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
pymongo==4.8.0
pytest==8.3.2
python-dotenv==1.0.1
sniffio==1.3.1
sqlparse==0.5.1
starlette==0.37.2
tomli==2.0.1
tqdm==4.66.5
typing_extensions==4.12.2
uvicorn==0.30.5
